Put SQLite connection pool behind Arc
Cloning connection pool causes database queries to fail when running
from different threads, because it seems that this actually causes the
database to be opened multiple times instead of accessing the single
connection through the pool.

This change puts the connection behind `Arc` so that cloning the pool
actually just clones the reference to the single pool instance.