Use PosgreSQL as the database
Combination of SQLite and Diesel is causing problems with concurrent
database queries and we need to migrate live environments away from it.

This change replaces use of SQLite with PosgreSQL, but doing so is
makes our tests very ugly and unstable.

This is just intermediate change while working towards better solution.