Move retry logic for HTTP requests into a function
Make it easier to reuse the retry logic for failed HTTP requests by
moving the logic into a separate function. For now, this is Reqwest
implementation specific one to simplify things, but later we should try
to create generic implementation that is then implemented by each HTTP
client integration. Currently it's just a bit tricky because the
implementation requires quite complex typing around futures and
lifetimes.